moving the yaml download logic to it's own component. Validation logic sits upstream of both submit and yaml download in a knowledge validation file. This will be moved up above skills too when that is refactored
