+++ src/components/Contribute/Knowledge/Submit/Submit.tsx
@@ -3,6 +3,7 @@ import { Button } from '@patternfly/react-core/dist/dynamic/components/Button';
 import { ActionGroupAlertContent, KnowledgeFormData, SeedExample } from '..';
 import { AttributionData, KnowledgeYamlData, SchemaVersion } from '@/types';
 import { dumpYaml } from '@/utils/yamlConfig';
+import { validateFields } from '../validation';
 
 interface Props {
   knowledgeFormData: KnowledgeFormData;
@@ -12,75 +13,6 @@ interface Props {
 }
 
 // temporary location of these validation functions. Once the Skills form has been refactored then these can be moved out to the utils file.
-export const validateEmail = (email: string): boolean => {
-  const emailRegex = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,6}$/;
-  return emailRegex.test(email);
-};
-
-export const hasDuplicateSeedExamples = (seedExamples: SeedExample[]): boolean => {
-  // Just checking contexts for duplication.
-  const contexts = new Set();
-
-  seedExamples.forEach((seedExample) => {
-    if (!contexts.has(seedExample.context)) {
-      contexts.add(seedExample.context);
-    } else {
-      return true;
-    }
-  });
-
-  return false;
-};
-
-export const validateFields = (
-  knowledgeFormData: KnowledgeFormData,
-  setActionGroupAlertContent: React.Dispatch<React.SetStateAction<ActionGroupAlertContent | undefined>>
-): boolean => {
-  // validate that data has been entered into all fields
-  for (const [key, value] of Object.entries(knowledgeFormData)) {
-    if (value === undefined) {
-      const actionGroupAlertContent: ActionGroupAlertContent = {
-        title: `Please make sure you complete the ${key} field`,
-        message: `Some fields are not filled out`,
-        success: false
-      };
-      setActionGroupAlertContent(actionGroupAlertContent);
-      return false;
-    }
-  }
-
-  //   Validate email
-
-  if (!validateEmail(knowledgeFormData.email!)) {
-    const actionGroupAlertContent: ActionGroupAlertContent = {
-      title: `Email address issue!`,
-      message: `Please enter a valid email address.`,
-      success: false
-    };
-    setActionGroupAlertContent(actionGroupAlertContent);
-    return false;
-  }
-
-  //   checking for seedExample duplication
-
-  if (hasDuplicateSeedExamples(knowledgeFormData.seedExamples)) {
-    const actionGroupAlertContent: ActionGroupAlertContent = {
-      title: `Seed example issue!`,
-      message: `There is duplicated context. Please provide unique contexts`,
-      success: false
-    };
-    setActionGroupAlertContent(actionGroupAlertContent);
-    return false;
-  }
-
-  const actionGroupAlertContent: ActionGroupAlertContent = {
-    title: `Data entry success`,
-    message: `All fields completed successfully`,
-    success: true
-  };
-  setActionGroupAlertContent(actionGroupAlertContent);
-  return true;
-};
 
 const Submit: React.FC<Props> = ({ knowledgeFormData, setActionGroupAlertContent, githubUsername, resetForm }) => {
   const handleSubmit = async (event: React.FormEvent<HTMLButtonElement>) => {
